requirements.txt
# Core dependencies
pyyaml>=6.0
tqdm>=4.65.0

# Audio transcription
faster-whisper>=0.9.0

# Audio playback (optional, used in mp3.py)
pygame>=2.5.0

# Translation services
deep-translator>=1.11.4
requests>=2.31.0
